Modified the forward of relu_conditional_transform to be more freindly with chaining together
For some reason I had the inverse_exp transform at the end of the transform, but if we chain mades together, we only call this transform at the end. I therefore moved it to the loss function. Note that I will need to store the logabsdet as a state variable for the final loss function calculateion when we chain many of these together

Forgot to add this change to previous commit